Update to rpi_lighting_node
Updates to fix incorrect switch statement in rpi_lighting_node

Also fixed naming conventions.
