+++ master_control/src/ilocator_python_lib/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="2">
-  <name>ilocator_pythonLib</name>
+  <name>ilocator_python_lib</name>
   <version>0.0.0</version>
-  <description>The ilocator_pythonLib package</description>
+  <description>The ilocator_python_lib package</description>
 
   <!-- One maintainer tag required, multiple allowed, one person per tag -->
   <!-- Example:  -->
