+++ master_control/src/ilocator_python_lib/setup.py
@@ -2,7 +2,7 @@ from distutils.core import setup
 from catkin_pkg.python_setup import generate_distutils_setup
 
 setup_args = generate_distutils_setup(
-    packages = ['ilocator_pythonLib'],
+    packages = ['ilocator_python_lib'],
     package_dir = {'': 'src'},
 )
