+++ master_control/src/rpi_lighting_node/CMakeLists.txt
